+++ src/Application/TodoLists/Commands/DeleteTodoList/DeleteTodoList.cs
@@ -4,9 +4,10 @@ namespace Stride.Application.TodoLists.Commands.DeleteTodoList;
 
 public record DeleteTodoListCommand(int Id) : IRequest;
 
-public class DeleteTodoListCommandHandler(IApplicationDbContext context) : IRequestHandler<DeleteTodoListCommand>
+public class DeleteTodoListCommandHandler(IApplicationDbContext context, IUser user) : IRequestHandler<DeleteTodoListCommand>
 {
     private readonly IApplicationDbContext _context = context;
+    private readonly IUser _user = user;
 
     public async Task Handle(DeleteTodoListCommand request, CancellationToken cancellationToken)
     {
@@ -15,6 +16,18 @@ public class DeleteTodoListCommandHandler(IApplicationDbContext context) : IRequ
             .SingleOrDefaultAsync(cancellationToken);
 
         Guard.Against.NotFound(request.Id, entity);
+
+        Guard.Against.Null(_user.Id);
+
+        var currentUser = await _context.StrideUsers
+            .FindAsync(new object[] { _user.Id }, cancellationToken);
+
+        Guard.Against.NotFound(_user.Id, currentUser);
+
+        if(entity.Owner != currentUser)
+        {
+            throw new UnauthorizedAccessException("User should be owner of the list.");
+        }
 
         _context.TodoLists.Remove(entity);
